Move file pattern matching to internal package without api
