+++ subprojects/files/src/main/java/org/gradle/internal/file/pattern/GreedyPathMatcher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2014 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,18 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.file.pattern;
+package org.gradle.internal.file.pattern;
 
-public class AnythingMatcher implements PathMatcher {
+public class GreedyPathMatcher implements PathMatcher {
+    private final PathMatcher next;
+
+    public GreedyPathMatcher(PathMatcher next) {
+        this.next = next;
+    }
+
     @Override
     public String toString() {
-        return "{anything}";
+        return "{greedy next: " + next + "}";
     }
 
     @Override
@@ -29,12 +35,19 @@ public class AnythingMatcher implements PathMatcher {
 
     @Override
     public int getMinSegments() {
-        return 0;
+        return next.getMinSegments();
     }
 
     @Override
     public boolean matches(String[] segments, int startIndex) {
-        return true;
+        int pos = segments.length - next.getMinSegments();
+        int minPos = Math.max(startIndex, segments.length - next.getMaxSegments());
+        for (; pos >= minPos; pos--) {
+            if (next.matches(segments, pos)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     @Override
