+++ subprojects/files/src/main/java/org/gradle/internal/file/pattern/PatternMatcher.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.file.pattern;
+package org.gradle.internal.file.pattern;
 
 @FunctionalInterface
 public interface PatternMatcher {
