+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/codegen/ApiExtensionsGenerator.kt
@@ -19,7 +19,7 @@
 package org.gradle.kotlin.dsl.codegen
 
 import org.gradle.api.file.RelativePath
-import org.gradle.api.internal.file.pattern.PatternMatcher
+import org.gradle.internal.file.pattern.PatternMatcher
 import org.gradle.kotlin.dsl.support.appendReproducibleNewLine
 import org.gradle.kotlin.dsl.support.useToRun
 import java.io.File
